requirements.txt
boto3
uv
kafka-python
minio
s3fs